requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
langgraph>=0.0.40
google-cloud-aiplatform>=1.43.0
google-cloud-logging>=3.8.0
google-cloud-translate>=3.12.0
google-cloud-firestore>=2.14.0
google-cloud-storage>=2.13.0
python-dotenv>=1.0.0
slowapi>=0.1.9
httpx>=0.26.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
